requirements.txt
datasets
torch
torchvision
tqdm
huggingface_hub
transformers
nltk
Pillow
matplotlib
python-dotenv